move shared functions to FeatureFilterUtils
